requirements.txt
mediapipe
opencv-python
scikit-learn
pandas
numpy
playsound
PyQt5
mysql-connector-python
matplotlib
